IDs return a full GraphQL type, not just a TypeName
This does no real harm at the protocol level and allows it to be used
for more things. For example, this gives us the option to use IDs as
cache keys not only for queries that return objects, but for queries
that return any value, whether or not that's a good idea.

We may want to only cache Objects if we want individual "function calls"
to not be cached by default, but it may also be the case that that
should just be done by marking the field @impure. In the latter case
(which sounds obvious now that I type it out), we'll want to be able to
use IDs as cache keys for all values, not just Objects.